requirements.txt
selenium
requests
googlesearch-python
nest_asyncio
aiohttp